Reorganize contents for forms package
